+++ play_lc/src/LC_E_0009.cpp
@@ -5,7 +5,7 @@ namespace lc
     int removeElement(std::vector<int>& nums, int val)
     {
         int index = 0;
-        for( int i = 0 ; i < nums.size();  i ++ )
+        for( size_t i = 0 ; i < nums.size();  i ++ )
             if( nums[i] != val )
                 nums[index++] = nums[i];
