+++ play_lc/src/LC_E_0012.cpp
@@ -5,12 +5,12 @@ namespace lc
     std::string countAndSay(int n)
     {
         std::string result = "1";
-        for( int x = 2; x <= n; x ++ )
+        for( size_t x = 2; x <= (size_t)n; x ++ )
         {
             std::string r = "";
             int     count = 1;
 
-            for (int i = 0; i < result.length(); ++i)
+            for (size_t i = 0; i < result.length(); ++i)
             {
                 if( i == (result.length() -1))
                 {
