+++ play_lc_once_again_top_google_questions/tests/TEST_merge_intervals.cpp
@@ -25,7 +25,7 @@ BOOST_AUTO_TEST_CASE( Case_1 )
 
 	auto res = lc::merge(intervals);
 
-	for ( int i = 0 ; i < res.size() ; ++ i)
+	for ( size_t i = 0 ; i < res.size() ; ++ i)
 	{
 		BOOST_CHECK(res[i][0] == result[i][0]);
 		BOOST_CHECK(res[i][1] == result[i][1]);
@@ -49,7 +49,7 @@ BOOST_AUTO_TEST_CASE( Case_2 )
 
 	auto res = lc::merge(intervals);
 
-	for ( int i = 0 ; i < res.size() ; ++ i)
+	for ( size_t i = 0 ; i < res.size() ; ++ i)
 	{
 		BOOST_CHECK(res[i][0] == result[i][0]);
 		BOOST_CHECK(res[i][1] == result[i][1]);
@@ -74,7 +74,7 @@ BOOST_AUTO_TEST_CASE( Case_3 )
 
 	auto res = lc::merge(intervals);
 
-	for ( int i = 0 ; i < res.size() ; ++ i)
+	for ( size_t i = 0 ; i < res.size() ; ++ i)
 	{
 		BOOST_CHECK(res[i][0] == result[i][0]);
 		BOOST_CHECK(res[i][1] == result[i][1]);
